tests: move inspect tests into dedicated test_inspect.py file
Collect all tests related to stdlib `inspect` module into a
dedicated `test_inspect.py` file.
